# docs/swagger.yaml
basePath: /
definitions:
...
  models.EstatisticasProdutos:
    properties:
      cavaletes_disponiveis:
        type: integer
      produtos_destaque:
        type: integer
      produtos_visiveis:
        type: integer
      total_produtos:
        type: integer
    type: object
...
  models.ProdutoAprovado:
    properties:
      cavalete_id:
        type: string
      created_at:
        type: string
      descricao:
        type: string
      destaque:
        type: boolean
      id:
        type: string
      nome_customizado:
        maxLength: 255
        minLength: 1
        type: string
      ordem_exibicao:
        type: integer
      preco_venda:
        type: number
      trader_id:
        type: string
      updated_at:
        type: string
      visivel:
        type: boolean
    required:
    - nome_customizado
    - preco_venda
    type: object
  models.ProdutoAprovarRequest:
    properties:
      cavalete_id:
        type: string
      descricao:
        type: string
      destaque:
        type: boolean
      nome_customizado:
        maxLength: 255
        minLength: 1
        type: string
      preco_venda:
        type: number
      visivel:
        type: boolean
    required:
    - cavalete_id
    - nome_customizado
    - preco_venda
    type: object
  models.ProdutoAtualizarRequest:
    properties:
      descricao:
        type: string
      destaque:
        type: boolean
      nome_customizado:
        maxLength: 255
        minLength: 1
        type: string
      ordem_exibicao:
        type: integer
      preco_venda:
        type: number
      visivel:
        type: boolean
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: "API para importação de ofertas do Mobgran com PostgreSQL - HOT RELOAD
    FUNCIONANDO! \U0001F525"
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Mobgran Importer API
  version: "1.0"
paths:
  /admin/limpar-dados:
    delete:
      description: Remove todos os registros de produtos, cavaletes, ofertas e dados
        relacionados do banco de dados
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Limpar todos os registros do banco de dados
      tags:
      - admin
...
  /produtos/{id}:
    delete:
      description: Remove um produto aprovado
      parameters:
      - description: ID do produto
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Remover produto
      tags:
      - produtos
    get:
      description: Busca um produto específico do trader
      parameters:
      - description: ID do produto
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ProdutoAprovado'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Buscar produto
      tags:
      - produtos
    put:
      consumes:
      - application/json
      description: Atualiza um produto aprovado
      parameters:
      - description: ID do produto
        in: path
        name: id
        required: true
        type: string
      - description: Dados para atualização
        in: body
        name: produto
        required: true
        schema:
          $ref: '#/definitions/models.ProdutoAtualizarRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ProdutoAprovado'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Atualizar produto
      tags:
      - produtos
  /produtos/aprovados:
    get:
      description: Lista produtos aprovados pelo trader
      parameters:
      - default: 20
        description: Limite de resultados
        in: query
        name: limit
        type: integer
      - default: 0
        description: Offset para paginação
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Listar produtos aprovados
      tags:
      - produtos
  /produtos/aprovar:
    post:
      consumes:
      - application/json
      description: Aprova um cavalete como produto para venda
      parameters:
      - description: Dados do produto
        in: body
        name: produto
        required: true
        schema:
          $ref: '#/definitions/models.ProdutoAprovarRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.ProdutoAprovado'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Aprovar produto
      tags:
      - produtos
  /produtos/cavaletes:
    get:
      description: Lista cavaletes disponíveis para aprovação pelo trader
      parameters:
      - default: 20
        description: Limite de resultados
        in: query
        name: limit
        type: integer
      - default: 0
        description: Offset para paginação
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Listar cavaletes disponíveis
      tags:
      - produtos
  /produtos/estatisticas:
    get:
      description: Obtém estatísticas dos produtos do trader
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.EstatisticasProdutos'
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Estatísticas de produtos
      tags:
      - produtos
...
  /vitrine/publica:
    get:
      description: Lista produtos na vitrine pública (não requer autenticação)
      parameters:
      - default: 20
        description: Limite de resultados
        in: query
        name: limit
        type: integer
      - default: 0
        description: Offset para paginação
        in: query
        name: offset
        type: integer
      - description: Filtrar por trader específico
        in: query
        name: trader_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Vitrine pública
      tags:
      - produtos
schemes:
- http
- https
swagger: "2.0"